requirements.txt
PyYAML
pandas
requests
tabulate
pylint
